split matchers
