+++ comparers/http_method_matcher.go
@@ -1,0 +1,9 @@
+package comparers
+
+import (
+	"strings"
+)
+
+func methodMatches(expected, actual string) bool {
+	return strings.EqualFold(expected, actual)
+}
